third_party/WebKit/LayoutTests/http/tests/security/referrer-policy-subresource.html

<!DOCTYPE html>
<html>
<head>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <meta name="referrer" content="no-referrer">
</head>
<body>
    <script src="resources/referrer-policy-script.php"></script>
    <script>
        // Test that referrer policies are applied on subresources.
        // crbug.com/508310
        test(function () {
            assert_equals("none", scriptReferrer);
        }, "Meta referrer policy is set on subresource");
    </script>
</body>
</html>
